+++ entrypoints/newtab/components/SettingsPage/Settings/MoreAbout.vue
@@ -24,26 +24,30 @@ const year = useDateFormat(useNow(), 'YYYY')
         class="settings__ext-link"
         href="https://github.com/Redlnn/lemon-new-tab-page/"
         target="_blank"
-        >{{ i18n.t('newtab.settings.about.projectLink') }}</a
       >
+        {{ i18n.t('newtab.settings.about.projectLink') }}
+      </a>
       <a
         class="settings__ext-link"
         href="https://chromewebstore.google.com/detail/bhbpmpflnpnkjanfgbjjhldccbckjohb"
         target="_blank"
-        >{{ i18n.t('newtab.settings.about.chromeExtensionLink') }}</a
       >
+        {{ i18n.t('newtab.settings.about.chromeExtensionLink') }}
+      </a>
       <a
         class="settings__ext-link"
         href="https://microsoftedge.microsoft.com/addons/detail/keikkgfgidagjlicckkangkfgnbdjdnh"
         target="_blank"
-        >{{ i18n.t('newtab.settings.about.edgeExtensionLink') }}</a
       >
+        {{ i18n.t('newtab.settings.about.edgeExtensionLink') }}
+      </a>
       <a
         class="settings__ext-link"
         href="https://addons.mozilla.org/zh-CN/firefox/addon/%E6%9F%A0%E6%AA%AC%E8%B5%B7%E5%A7%8B%E9%A1%B5/"
         target="_blank"
-        >{{ i18n.t('newtab.settings.about.firefoxExtensionLink') }}</a
       >
+        {{ i18n.t('newtab.settings.about.firefoxExtensionLink') }}
+      </a>
     </div>
     <div class="settings__ext-links">
       <span>