+++ shared/settings/current.ts
@@ -1,4 +1,4 @@
-import type { SettingsInterfaceVer5 } from './types'
+import type { SettingsInterfaceVer6 } from './types'
 
-export const CURRENT_CONFIG_VERSION = 5
-export type CURRENT_CONFTG_INTERFACE = SettingsInterfaceVer5
+export const CURRENT_CONFIG_VERSION = 6
+export type CURRENT_CONFIG_INTERFACE = SettingsInterfaceVer6