+++ shared/settings/index.ts
@@ -1,16 +1,18 @@
 export { BgType } from './types'
 export {
-  migrateFromVer1To5,
+  migrateFromVer1,
   migrateFromVer2To3,
   migrateFromVer3To4,
-  migrateFromVer4To5
+  migrateFromVer4To5,
+  migrateFromVer5To6
 } from './migrate'
 export type {
   OldSettingsInterface,
   SettingsInterfaceVer2,
   SettingsInterfaceVer3,
   SettingsInterfaceVer4,
-  SettingsInterfaceVer5
+  SettingsInterfaceVer5,
+  SettingsInterfaceVer6
 } from './types'
 export {
   initSettings,
@@ -21,5 +23,5 @@ export {
 } from './settingsStore'
 export { settingsStorage } from './settingsStorage'
 export { useBingWallpaperStore, useWallpaperStore } from './wallpaperStore'
-export { CURRENT_CONFIG_VERSION, type CURRENT_CONFTG_INTERFACE } from './current'
+export { CURRENT_CONFIG_VERSION, type CURRENT_CONFIG_INTERFACE } from './current'
 export { defaultSettings } from './default'